api con dang loi vai cai thieu end-date de mai t cap nhat tiep

+++ src/main/java/com/devsoga/BookStore_V2/repositories/InventoryRepository.java
@@ -3,5 +3,5 @@ package com.devsoga.BookStore_V2.repositories;
 import com.devsoga.BookStore_V2.enties.InventoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InventoryRepository extends JpaRepository<InventoryEntity, Long> {
+public interface InventoryRepository extends JpaRepository<InventoryEntity, Integer> {
 }
