+++ src/main/java/com/devsoga/BookStore_V2/repositories/InvoiceDetailRepository.java
@@ -3,5 +3,5 @@ package com.devsoga.BookStore_V2.repositories;
 import com.devsoga.BookStore_V2.enties.InvoiceDetailEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InvoiceDetailRepository extends JpaRepository<InvoiceDetailEntity, Long> {
+public interface InvoiceDetailRepository extends JpaRepository<InvoiceDetailEntity, Integer> {
 }
