+++ src/main/java/com/devsoga/BookStore_V2/repositories/InvoiceRepository.java
@@ -3,5 +3,5 @@ package com.devsoga.BookStore_V2.repositories;
 import com.devsoga.BookStore_V2.enties.InvoiceEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InvoiceRepository extends JpaRepository<InvoiceEntity, Long> {
+public interface InvoiceRepository extends JpaRepository<InvoiceEntity, Integer> {
 }
