+++ src/main/java/com/devsoga/BookStore_V2/repositories/PromotionRepository.java
@@ -3,5 +3,5 @@ package com.devsoga.BookStore_V2.repositories;
 import com.devsoga.BookStore_V2.enties.PromotionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PromotionRepository extends JpaRepository<PromotionEntity, Long> {
+public interface PromotionRepository extends JpaRepository<PromotionEntity, Integer> {
 }
