+++ src/main/java/com/devsoga/BookStore_V2/repositories/ReturnOrderRepository.java
@@ -3,5 +3,5 @@ package com.devsoga.BookStore_V2.repositories;
 import com.devsoga.BookStore_V2.enties.ReturnOrderEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReturnOrderRepository extends JpaRepository<ReturnOrderEntity, Long> {
+public interface ReturnOrderRepository extends JpaRepository<ReturnOrderEntity, Integer> {
 }
